Update tool_2.py
Lines 113-124:  Added a check to see if a buffer file already exists and, if so, move directly to the clip operation.  Otherwise, the buffer file is created then waypoints clipped.  
Lines 198-199:  Added a message at the end communicating the number of aircraft flights having a null N-number in addition to the tool execution time.